requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.1
django-pwa==2.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
